+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-21 18:12:11
-// Commit: fix(test): correct layout constraints and overflow in test
+// Updated at: 2026-08-23 09:52:21
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '3800';
-  final String lastMessage = 'fix(test): correct layout constraints and overflow in test';
-  final DateTime timestamp = DateTime.parse('2026-08-21T18:12:11+05:00');
+  final String commitRef = '3822';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-08-23T09:52:21+05:00');
 }